Make a speech-complete timeout configurable
Before this change there was no timeout for sending final recognition
results. This change adds a configurable timeout (default is 0
seconds). The final recognition result a concatenation of all
intermediate "final" recognition results, and the final confidence
score is an their average. The periods at the ends of sentences are
removed. Pause button resets the result and recognition will start
over. Additionally, the change adds some code for testing the widget
internally.