Add external event for closing the widget
This change adds *talaStop* CustomEvent. When received, all state
charts go to 'stopped' state. The intended behaviour is that the
widget should be closed after that. In order to start it again, the
page can either be reloaded or *talaClick* event can be sent. The
widget will go to its initial state after the click.

Additionally, in this change formats the code according to the
https://www.npmjs.com/package/prettier.