Wait infinitely if passivity is null
Before this change, tala-speech was resetting TDM passivity value
equal to null to default hard-coded value. Now it will wait
infinitely (technically, it is 24 hours).

The default value is now used in only one case -- when user pauses the
recognition and then starts it again. In this case we assume that the
user is about to say something, and we don't need to wait for too
long.