Make sure streaming isn't aborted before end
Previously, the condition for moving out of the buffered speaker state
was that the TTS stopped speaking and that the buffer was empty. This is
unsafe, since the buffer can be empty when the backend is still
streaming an utterance to the frontend.

This change makes sure that the buffer is empty, and that a DONE event
has happened before drawing the conclusion that the stream is ended.