Lightweight UI for tala-speech
This change adds a UI interface which can be rendered using
`window.TalaSpeechRenderer`. This object contains 2 functions:

* `renderTalaSpeech`: takes 3 arguments: `settings, page, element`;
renders UI button inside the `element`, starts session with the
`settings` provided, goes to a DDD defined by `page`, and prepares
SpeechState.

* `getDialogueJson` is a helper function which fetches the JSON given
the URL for dialoguejson API.

The example of usage can be found in `index.html`.

In addition, this change switches build and development processes to
`esbuild` instead `vite`.

CSS styling is streamlined using Tailwind.